password test to own file
